removed `network_utilities`
As mention previously, whole `network` stuff was moved entirely to the server, so the utils are no longer used. I also noticed that I was unintentionally collect (not store is db, just in variable) user's ip address. So it was fixed as well. There is no longer any IP mention in the entire code for that moment (I'm never planning to add it)