fixed `reset` & `initialize` every security layer
Previous implementation was full of bugs when user tried to switch between views in app or view in tabs (in settings view). It leaded to inproper resetting (incorrect re-rendering) at every layer due to opacity animation. This implementation fixed that problem and is properly handling all of the switches between views & tabs and also buttons: back and restore to default.